Laid Foundations for PieceBuilder
Added the PieceBuilder, which allows for custom pieces to be created. It
supports custom images which can be drag-and-dropped into the builder.
It currently supports all action types except MultiActions, and does not
yet support Conditions, Child Nodes, or Choke nodes (bottleneck points).
These will be added in future commits.